requirements.txt
python_jenkins==1.8.1
